requirements.txt
dj-database-url==0.4.2
Django==1.11.5
django-crispy-forms==1.6.1
django-markdown-deux==1.0.5
django-pagedown==0.1.3
gunicorn==19.7.1
markdown2==2.3.4
olefile==0.44
Pillow==4.2.1
psycopg2==2.7.3.1
pytz==2017.2
